fix a few bugs and cleanup some code
