+++ src/index.ts
@@ -23,5 +23,3 @@ async function startServer() {
 }
 
 startServer();
-
-console.log("index");
